requirements.txt
django==1.5
django-tastypie>=0.9.12
django-templatetag-handlebars>=1.2.0
django-openid-auth
python-openid
